Improve safety of image unmount and output errors.
Previously unmount failures were totally repressed often leaving loop devices
and mounts in place, unexpectedly.

We wanted to avoid the 'not mounted' error on unmount so we can use as a safety
mechanism to ensure (as best we can) that images are unmounted on error in other
script functions. If this outputs an error then the whole script dies and this
doesn't work.

To workaround, simply check to make sure the device is mounted before attempting
an unmount. Now we can output errors and stop if something goes wrong.