Switch to using a debian rootfs.
I hadn't realised that the debian archive actually had packages available! This
combined with the awesome debootstrap makes building a useable image a LOT
easier.

This still needs some more work but the image is already a lot more useable and
helps me avoid compiling binutils/glibc/gcc which is always going to put me in a
good mood!

I will work on networking so we can take advantage of packages.